# perf(query): Update CompressedBin IntersectionAlgo